+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Remove redundant `beginBuild`/`endBuild` (formerly `start`/`end`) events on builder
+
 # 1.0.0-beta.4
 
 * Improve test suite